Редактирование cshtml документов

+++ Models/Electronic.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string? Name { get; set; } // имя товара
         public string? Category { get; set; } // категория товара
-        public float? Price { get; set; } // цена товара
+        public double? Price { get; set; } // цена товара
         public int ForSale { get; set; } // кол-во для продажи
         public int Sold { get; set; } // кол-во проданного товара
         public int Counter { get; set; } // счетчик проданных товаров
